Refactor leaflet methods into map-manager.ts
