+++ src/interface/src/app/map/map.component.ts
@@ -8,16 +8,6 @@ import {
   OnInit,
 } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import {
-  Feature,
-  FeatureCollection,
-  Geometry,
-  MultiPolygon,
-  Polygon,
-} from 'geojson';
-import * as L from 'leaflet';
-import 'leaflet-draw';
-import 'leaflet.sync';
 import { BehaviorSubject, Observable, Subject, take, takeUntil } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 
@@ -37,8 +27,8 @@ import {
   MapViewOptions,
   Region,
 } from '../types';
-import { BackendConstants } from '../backend-constants';
 import { Legend } from './../shared/legend/legend.component';
+import { MapManager } from './map-manager';
 import { PlanCreateDialogComponent } from './plan-create-dialog/plan-create-dialog.component';
 import { ProjectCardComponent } from './project-card/project-card.component';
 
@@ -53,6 +43,8 @@ export interface PlanCreationOption {
   styleUrls: ['./map.component.scss'],
 })
 export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
+  mapManager: MapManager;
+
   maps: Map[];
   mapViewOptions: MapViewOptions = {
     selectedMapIndex: 0,
@@ -106,29 +98,6 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
   ];
   selectedPlanCreationOption: PlanCreationOption | null = null;
   showCreatePlanButton: boolean = false;
-  drawingLayer = new L.FeatureGroup();
-
-  static hillshadeTiles() {
-    return L.tileLayer(
-      'https://api.mapbox.com/styles/v1/tsuga11/ckcng1sjp2kat1io3rv2croyl/tiles/{z}/{x}/{y}?access_token=pk.eyJ1IjoidHN1Z2ExMSIsImEiOiJjanFmaTA5cGIyaXFoM3hqd3R5dzd3bzU3In0.TFqMjIIYtpcyhzNh4iMcQA',
-      {
-        zIndex: 0,
-        tileSize: 512,
-        zoomOffset: -1,
-      }
-    );
-  }
-
-  static stadiaAlidadeTiles() {
-    return L.tileLayer(
-      'https://tiles.stadiamaps.com/tiles/alidade_smooth/{z}/{x}/{y}{r}.png',
-      {
-        maxZoom: 19,
-        attribution:
-          '&copy; <a href="https://stadiamaps.com/" target="_blank" rel="noreferrer">Stadia Maps</a>, &copy; <a href="https://openmaptiles.org/" target="_blank" rel="noreferrer">OpenMapTiles</a> &copy; <a href="http://openstreetmap.org" target="_blank" rel="noreferrer">OpenStreetMap</a> contributors',
-      }
-    );
-  }
 
   private readonly destroy$ = new Subject<void>();
 
@@ -220,6 +189,8 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
         };
       }
     );
+
+    this.mapManager = new MapManager(this.maps, popupService);
   }
 
   ngOnInit(): void {
@@ -240,9 +211,10 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
       this.initMap(map, map.id);
     });
 
-    this.syncAllMaps();
-
-    this.addDrawingControls(this.maps[0].instance!);
+    this.mapManager.syncAllMaps();
+    this.mapManager.addDrawingControls(this.maps[0].instance!, () => {
+      this.showCreatePlanButton = true;
+    });
   }
 
   ngOnDestroy(): void {
@@ -272,160 +244,35 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
   }
 
   /** Initializes the map with controls and the layer options specified in its config. */
-  private initMap(map: Map, id: string): L.Map {
-    if (map.instance != undefined) map.instance.remove();
-
-    if (map.config.baseLayerType === BaseLayerType.Road) {
-      map.baseLayerRef = MapComponent.stadiaAlidadeTiles();
-    } else {
-      map.baseLayerRef = MapComponent.hillshadeTiles();
-    }
-
-    map.instance = L.map(id, {
-      center: [38.646, -120.548],
-      zoom: 9,
-      layers: [map.baseLayerRef],
-      zoomControl: false,
-    });
-
-    // Add zoom controls to bottom right corner
-    const zoomControl = L.control.zoom({
-      position: 'bottomright',
-    });
-    zoomControl.addTo(map.instance);
-
-    // Init layers, but only add them to the map instance if specified in the map config.
-    this.huc12BoundaryGeoJson$.subscribe((boundary: GeoJSON.GeoJSON | null) => {
-      if (boundary) {
-        this.initHuc12BoundaryLayer(map, boundary);
-      }
-    });
-    this.huc10BoundaryGeoJson$.subscribe((boundary: GeoJSON.GeoJSON | null) => {
-      if (boundary != null) {
-        this.initHUC10BoundaryLayer(map, boundary);
-      }
-    });
-    this.countyBoundaryGeoJson$.subscribe(
-      (boundary: GeoJSON.GeoJSON | null) => {
-        if (boundary) {
-          this.initCountyBoundaryLayer(map, boundary);
-        }
-      }
-    );
-    this.usForestBoundaryGeoJson$.subscribe(
-      (boundary: GeoJSON.GeoJSON | null) => {
-        if (boundary != null) {
-          this.initUSForestBoundaryLayer(map, boundary);
-        }
-      }
-    );
-    this.existingProjectsGeoJson$.subscribe(
-      (projects: GeoJSON.GeoJSON | null) => {
-        if (projects) {
-          this.initCalMapperLayer(map, projects);
-        }
+  private initMap(map: Map, id: string) {
+    this.mapManager.initLeafletMap(
+      map,
+      id,
+      this.huc12BoundaryGeoJson$,
+      this.huc10BoundaryGeoJson$,
+      this.countyBoundaryGeoJson$,
+      this.usForestBoundaryGeoJson$,
+      this.existingProjectsGeoJson$,
+      (feature) => {
+        let component = createComponent(ProjectCardComponent, {
+          environmentInjector: this.environmentInjector,
+        });
+        component.instance.feature = feature;
+        this.applicationRef.attachView(component.hostView);
+        return component.location.nativeElement;
       }
     );
 
     // Renders the selected region on the map.
     this.selectedRegion$.subscribe((selectedRegion: Region | null) => {
-      this.displayRegionBoundary(map.instance!, selectedRegion);
+      this.displayRegionBoundary(map, selectedRegion);
     });
 
     // Mark the map as selected when the user clicks anywhere on it.
-    map.instance.addEventListener('click', () => {
+    map.instance?.addEventListener('click', () => {
       this.mapViewOptions.selectedMapIndex = this.maps.indexOf(map);
       this.sessionService.setMapViewOptions(this.mapViewOptions);
     });
-
-    return map.instance;
-  }
-
-  /** Sync pan, zoom, etc. between all maps. */
-  private syncAllMaps() {
-    this.maps.forEach((mapA) => {
-      this.maps.forEach((mapB) => {
-        if (mapA !== mapB) {
-          (mapA.instance as any).sync(mapB.instance);
-        }
-      });
-    });
-  }
-
-  /** Adds drawing controls and handles drawing events. */
-  private addDrawingControls(map: L.Map) {
-    map.addLayer(this.drawingLayer);
-
-    const drawOptions: L.Control.DrawConstructorOptions = {
-      position: 'bottomright',
-      draw: {
-        polygon: {
-          allowIntersection: false,
-          showArea: true,
-          metric: false, // Set measurement units to acres
-          shapeOptions: {
-            color: '#7b61ff',
-          },
-          drawError: {
-            color: '#ff7b61',
-            message: "Can't draw polygons with intersections!",
-          },
-        }, // Set to false to disable each tool
-        polyline: false,
-        circle: false,
-        rectangle: false,
-        marker: false,
-        circlemarker: false,
-      },
-      edit: {
-        featureGroup: this.drawingLayer, // Required and declares which layer is editable
-      },
-    };
-
-    const drawControl = new L.Control.Draw(drawOptions);
-    map.addControl(drawControl);
-
-    this.setUpDrawingHandlers(map);
-  }
-
-  private setUpDrawingHandlers(map: L.Map) {
-    map.on('draw:created', (event) => {
-      const layer = (event as L.DrawEvents.Created).layer;
-      this.drawingLayer.addLayer(layer);
-
-      this.showCreatePlanButton = true;
-    });
-  }
-
-  /**
-   * Converts drawingLayer to GeoJSON. If there are multiple polygons drawn,
-   * creates and returns MultiPolygon type GeoJSON. Otherwise, returns a Polygon
-   * type GeoJSON.
-   */
-  private convertToPlanningArea(): GeoJSON.GeoJSON {
-    const drawnGeoJson = this.drawingLayer.toGeoJSON() as FeatureCollection;
-    // Case: Single polygon
-    if (drawnGeoJson.features.length <= 1) return drawnGeoJson;
-
-    // Case: Multipolygon
-    const newFeature: GeoJSON.Feature = {
-      type: 'Feature',
-      geometry: {
-        type: 'MultiPolygon',
-        coordinates: [],
-      },
-      properties: {},
-    };
-    drawnGeoJson.features.forEach((feature) => {
-      (newFeature.geometry as MultiPolygon).coordinates.push(
-        (feature.geometry as Polygon).coordinates
-      );
-    });
-
-    return {
-      type: 'FeatureCollection',
-      features: [newFeature],
-    } as FeatureCollection;
   }
 
   /** Configures and opens the Create Plan dialog */
@@ -440,7 +287,7 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
 
     openedDialog.afterClosed().subscribe((result) => {
       if (result) {
-        this.createPlan(result.value, this.convertToPlanningArea());
+        this.createPlan(result.value, this.mapManager.convertToPlanningArea());
       }
     });
   }
@@ -468,264 +315,54 @@ export class MapComponent implements AfterViewInit, OnDestroy, OnInit {
    */
   onPlanCreationOptionChange(option: PlanCreationOption) {
     if (option.value === 'draw-area') {
-      const polygonDrawer = new L.Draw.Polygon(
-        this.maps[0].instance as L.DrawMap,
-        {
-          allowIntersection: false,
-          showArea: true,
-          metric: false,
-          shapeOptions: {
-            color: '#7b61ff',
-          },
-          drawError: {
-            color: '#ff7b61',
-            message: "Can't draw polygons with intersections!",
-          },
-        }
-      );
-      polygonDrawer.enable();
+      this.mapManager.enablePolygonDrawingTool(this.maps[0].instance!);
     }
   }
 
   /** Gets the selected region geojson and renders it on the map. */
-  private displayRegionBoundary(map: L.Map, selectedRegion: Region | null) {
+  private displayRegionBoundary(map: Map, selectedRegion: Region | null) {
     if (!selectedRegion) return;
+    if (!map.instance) return;
     this.mapService
       .getRegionBoundary(selectedRegion)
       .subscribe((boundary: GeoJSON.GeoJSON) => {
-        this.maskOutsideRegion(map, boundary);
+        this.mapManager.maskOutsideRegion(map.instance!, boundary);
       });
-  }
-
-  /**
-   * Darkens everything outside of the region boundary.
-   * Type 'any' is used in order to access coordinates.
-   */
-  private maskOutsideRegion(map: L.Map, boundary: any) {
-    // Add corners of the map to invert the polygon
-    boundary.features[0].geometry.coordinates[0].unshift([
-      [180, -90],
-      [180, 90],
-      [-180, 90],
-      [-180, -90],
-    ]);
-    L.geoJSON(boundary, {
-      style: (feature) => ({
-        color: '#ffffff',
-        weight: 2,
-        opacity: 1,
-        fillColor: '#000000',
-        fillOpacity: 0.4,
-      }),
-    }).addTo(map);
-  }
-
-  /** Renders the existing project boundaries + metadata in a popup in an optional layer. */
-  private initCalMapperLayer(map: Map, existingProjects: GeoJSON.GeoJSON) {
-    // [elsieling] This step makes the map less responsive
-    map.existingProjectsLayerRef = L.geoJSON(existingProjects, {
-      style: function (_) {
-        return {
-          color: '#000000',
-          weight: 3,
-          opacity: 0.9,
-        };
-      },
-      onEachFeature: (feature: Feature<Geometry, any>, layer: L.Layer) => {
-        let component = createComponent(ProjectCardComponent, {
-          environmentInjector: this.environmentInjector,
-        });
-        component.instance.feature = feature;
-        this.applicationRef.attachView(component.hostView);
-        layer.bindPopup(component.location.nativeElement);
-      },
-    });
-
-    if (map.config.showExistingProjectsLayer) {
-      map.instance?.addLayer(map.existingProjectsLayerRef);
-    }
-  }
-
-  /** Renders the HUC-12 boundaries in an optional layer. */
-  private initHuc12BoundaryLayer(map: Map, boundary: GeoJSON.GeoJSON) {
-    map.huc12BoundaryLayerRef = L.geoJSON(boundary, {
-      style: (_) => ({
-        weight: 3,
-        opacity: 0.5,
-        color: '#0000ff',
-        fillOpacity: 0.2,
-        fillColor: '#6DB65B',
-      }),
-      onEachFeature: (feature, layer) =>
-        layer.bindPopup(
-          this.popupService.makeDetailsPopup(feature.properties.shape_name)
-        ),
-    });
-
-    if (map.config.showHuc12BoundaryLayer) {
-      map.instance?.addLayer(map.huc12BoundaryLayerRef);
-    }
-  }
-
-  private initHUC10BoundaryLayer(map: Map, boundary: GeoJSON.GeoJSON) {
-    map.huc10BoundaryLayerRef = L.geoJSON(boundary, {
-      style: (feature) => ({
-        weight: 3,
-        opacity: 0.5,
-        color: '#0000ff',
-        fillOpacity: 0.2,
-        fillColor: '#6DB65B',
-      }),
-      onEachFeature: (feature, layer) =>
-        layer.bindPopup(
-          this.popupService.makeDetailsPopup(feature.properties.shape_name)
-        ),
-    });
-
-    if (map.config.showHuc10BoundaryLayer) {
-      map.instance?.addLayer(map.huc10BoundaryLayerRef);
-    }
-  }
-
-  /** Renders the county boundaries in an optional layer. */
-  private initCountyBoundaryLayer(map: Map, boundary: GeoJSON.GeoJSON) {
-    map.countyBoundaryLayerRef = L.geoJSON(boundary, {
-      style: (feature) => ({
-        weight: 3,
-        opacity: 0.5,
-        color: '#0000ff',
-        fillOpacity: 0.2,
-        fillColor: '#6DB65B',
-      }),
-      onEachFeature: (feature, layer) =>
-        layer.bindPopup(
-          this.popupService.makeDetailsPopup(feature.properties.shape_name)
-        ),
-    });
-
-    if (map.config.showCountyBoundaryLayer) {
-      map.instance?.addLayer(map.countyBoundaryLayerRef);
-    }
-  }
-
-  private initUSForestBoundaryLayer(map: Map, boundary: GeoJSON.GeoJSON) {
-    map.usForestBoundaryLayerRef = L.geoJSON(boundary, {
-      style: (feature) => ({
-        weight: 3,
-        opacity: 0.5,
-        color: '#0000ff',
-        fillOpacity: 0.2,
-        fillColor: '#6DB65B',
-      }),
-      onEachFeature: (feature, layer) =>
-        layer.bindPopup(
-          this.popupService.makeDetailsPopup(feature.properties.shape_name)
-        ),
-    });
-
-    if (map.config.showUsForestBoundaryLayer) {
-      map.instance?.addLayer(map.usForestBoundaryLayerRef);
-    }
   }
 
   /** Toggles which base layer is shown. */
   changeBaseLayer(map: Map) {
-    let baseLayerType = map.config.baseLayerType;
-    map.baseLayerRef?.remove();
-    if (baseLayerType === BaseLayerType.Terrain) {
-      map.baseLayerRef = MapComponent.hillshadeTiles();
-    } else if (baseLayerType === BaseLayerType.Road) {
-      map.baseLayerRef = MapComponent.stadiaAlidadeTiles();
-    }
-    map.instance?.addLayer(map.baseLayerRef!);
+    this.mapManager.changeBaseLayer(map);
   }
 
   /** Toggles whether HUC-12 boundaries are shown. */
   toggleHuc12BoundariesLayer(map: Map) {
-    if (map.instance === undefined) return;
-
-    if (map.config.showHuc12BoundaryLayer) {
-      map.huc12BoundaryLayerRef?.addTo(map.instance);
-    } else {
-      map.huc12BoundaryLayerRef?.remove();
-    }
+    this.mapManager.toggleHuc12BoundariesLayer(map);
   }
 
   /** Toggles whether HUC-10 boundaries are shown. */
   toggleHUC10BoundariesLayer(map: Map) {
-    if (map.instance === undefined) return;
-
-    if (map.config.showHuc10BoundaryLayer) {
-      map.huc10BoundaryLayerRef?.addTo(map.instance);
-    } else {
-      map.huc10BoundaryLayerRef?.remove();
-    }
+    this.mapManager.toggleHUC10BoundariesLayer(map);
   }
 
   /** Toggles whether county boundaries are shown. */
   toggleCountyBoundariesLayer(map: Map) {
-    if (map.instance === undefined) return;
-
-    if (map.config.showCountyBoundaryLayer) {
-      map.countyBoundaryLayerRef?.addTo(map.instance);
-    } else {
-      map.countyBoundaryLayerRef?.remove();
-    }
+    this.mapManager.toggleCountyBoundariesLayer(map);
   }
 
   /** Toggles whether US Forest boundaries are shown. */
   toggleUSForestsBoundariesLayer(map: Map) {
-    if (map.instance == undefined) return;
-
-    if (map.config.showUsForestBoundaryLayer) {
-      map.usForestBoundaryLayerRef?.addTo(map.instance);
-    } else {
-      map.usForestBoundaryLayerRef?.remove();
-    }
+    this.mapManager.toggleUSForestsBoundariesLayer(map);
   }
 
   /** Toggles whether existing projects from CalMapper are shown. */
   toggleExistingProjectsLayer(map: Map) {
-    if (map.instance === undefined) return;
-
-    if (map.config.showExistingProjectsLayer) {
-      map.existingProjectsLayerRef?.addTo(map.instance);
-    } else {
-      map.existingProjectsLayerRef?.remove();
-    }
+    this.mapManager.toggleExistingProjectsLayer(map);
   }
 
   /** Changes which condition scores layer (if any) is shown. */
   changeConditionsLayer(map: Map) {
-    if (map.instance === undefined) return;
-
-    map.dataLayerRef?.remove();
-
-    let filepath = map.config.dataLayerConfig.filepath;
-    if (filepath?.length === 0 || !filepath) return;
-    if (map.config.normalizeDataLayer) {
-      filepath = filepath.concat('_normalized');
-    }
-    filepath = filepath.substring(filepath.lastIndexOf('/') + 1) + '.tif';
-
-    let colormap = map.config.dataLayerConfig.colormap;
-    if (colormap?.length === 0 || !colormap) {
-      colormap = 'viridis';
-    }
-
-    map.dataLayerRef = L.tileLayer.wms(BackendConstants.END_POINT + '/conditions/wms', {
-      crs: L.CRS.EPSG4326,
-      minZoom: 7,
-      maxZoom: 15,
-      format: 'image/png',
-      opacity: 0.7,
-      layers: filepath,
-      styles: colormap,
-    });
-
-    if (map.config.showDataLayer) {
-      map.dataLayerRef.addTo(map.instance);
-    }
+    this.mapManager.changeConditionsLayer(map);
   }
 
   /* Change how many maps are displayed in the viewport. */
